Only cache `gulp links` in CI

+++ src/tasks/links/main.js
@@ -1,6 +1,7 @@
 import { writeFile } from 'node:fs/promises'
 
 import { exec } from 'gulp-execa'
+import isCi from 'is-ci'
 import { pathExists } from 'path-exists'
 
 import { LINKS_CACHE_FILE } from '../../files.js'
@@ -12,7 +13,7 @@ import { getExcludedLinks, getRemaps } from './urls.js'
 
 // Detect dead links with lychee
 export const links = async () => {
-  if (!(await pathExists(LINKS_CACHE_FILE))) {
+  if (isCi && !(await pathExists(LINKS_CACHE_FILE))) {
     await writeFile(LINKS_CACHE_FILE, '')
   }
 
@@ -30,8 +31,7 @@ export const links = async () => {
 }
 
 const MAIN_FLAGS = [
-  '--cache',
-  '--max-cache-age=1h',
+  ...(isCi ? ['--cache', '--max-cache-age=1h'] : []),
   '--include-fragments',
   '--verbose',
   '--format=detailed',
